Baseline the giraffe and tree, and Container refactor.
Wanting to baseline the giraffe and tree by changing the anchor point
necessitated a refactor to use Container's to group the sprites for
each object. This implies that I should probably directly expose the
parent object so that anchor, rotation etc can be played with.

I got a little bit tricked here as at first I tried to resize the
container's to match their child element's dimensions, and didn't
realise that `getBounds` is the function to return the **global**
bounds. Seems that you should just leave container's as a 16x16 unless
you want to scale their children, regardless of how big the child's are.