Optimise following guidance here https://developer.chrome.com/docs/devtools/evaluate-performance/.
Turned out that the biggest expenses were calls to getBounds, which was
a surprise, and the quadratic giraffe/apple collision detector, which
wasn't a surprise.

I addressed the getBounds calls through memoization on each tick, so at
least it's not getting re-calculated multiple times per object per tick,
and addressed the quadratic algorithm by making it quadratic in
giraffe/tree instead of giraffe/apple, a significant improvement as
there is a fixed, small number of both giraffes and trees.